Arduino code for Elysa 3.0
I need to pull the latest version of the GUI out of the robot, I will upload that later.

Sorry if I have provided any outdated software versions here, If I realise I have done this I will replace them

GUI sends serial commands to the arduino mega.
Arduino mega handles driving and sensor feedback
Tranceiver board is not complete, but it handles the key switch. It also controls the front panel buttons and reads the inturnal temprature/humidity. 
On boot each board is contacted to see if its alive, results are displayed on screen
Left arm and right arm are controlling the linear extension of each side
head controls the PWM for the neck servos and projector deployment servo along with SPI for the dot matrix
MUX board is a multiplexer that turns on/off the leds on screen. 

Mega is a i2c master. it also reads the voltage, humitiy/temp exturnally
I2c slaves include all other arduinos, two LCDs, IMU and a PWM generator for the rest of the servos.